Module chess_gui::slint_generatedBoard_UI

Structs

Board_UI

InnerBoard_UI

InnerBoard_UIFieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerBoard_UI

InnerColorSchemeSelector_11 🔒

InnerColorSchemeSelector_11FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerColorSchemeSelector_11

InnerComponent_empty_4 🔒

InnerComponent_empty_4FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_empty_4

InnerComponent_square_5 🔒

InnerComponent_square_5FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_square_5

InnerFluentPalette_12 🔒

InnerFluentPalette_12FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerFluentPalette_12

InnerPieceImg_root_10

InnerPieceImg_root_10FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerPieceImg_root_10

Piece_UI

SharedGlobals 🔒

Enums

PieceColour_UI

PieceType_UI

Constants

_THE_SAME_VERSION_MUST_BE_USED_FOR_THE_COMPILER_AND_THE_RUNTIME 🔒

Statics

SLINT_EMBEDDED_RESOURCE_0 🔒

SLINT_EMBEDDED_RESOURCE_1 🔒

SLINT_EMBEDDED_RESOURCE_2 🔒

SLINT_EMBEDDED_RESOURCE_3 🔒

SLINT_EMBEDDED_RESOURCE_4 🔒

SLINT_EMBEDDED_RESOURCE_5 🔒

SLINT_EMBEDDED_RESOURCE_6 🔒

SLINT_EMBEDDED_RESOURCE_7 🔒

SLINT_EMBEDDED_RESOURCE_8 🔒

SLINT_EMBEDDED_RESOURCE_9 🔒

SLINT_EMBEDDED_RESOURCE_10 🔒

SLINT_EMBEDDED_RESOURCE_11 🔒